+++ src/providers/sfdc-text-document-provider.ts
@@ -1,6 +1,9 @@
 import * as jsforce from 'jsforce';
+import * as querystring from 'querystring';
 import { TextDocumentContentProvider, EventEmitter, Uri, Event } from 'vscode';
 import { queryRecordsById } from '../common/sfdc-utils';
+import * as _ from 'lodash';
+import { CustomScript } from '../models';
 
 export class SfdcTextDocumentProvider implements TextDocumentContentProvider {
   private _onDidChange = new EventEmitter<Uri>();
@@ -13,8 +16,16 @@ export class SfdcTextDocumentProvider implements TextDocumentContentProvider {
 
   public async provideTextDocumentContent(uri: Uri): Promise<string> {
     const records = await queryRecordsById(this.conn, uri.fragment);
+    const queryString = querystring.parse(uri.query);
     if (records && records[0]) {
-      return records[0].SBQQ__Code__c;
+      if (_.isString(queryString['field'])) {
+        if (!(records[0] as any)[queryString['field'] as keyof CustomScript]) {
+          throw new Error(`The field ${queryString['field']} is blank.`);
+        }
+        return (records[0] as any)[queryString['field'] as keyof CustomScript];
+      } else {
+        return records[0].SBQQ__Code__c;
+      }
     }
     throw new Error(`A record with Id ${uri.path} was not found on Salesforce.`);
   }